refator : 데이터 받는 코드 변경, 화면의 사용성 고려한 수정, MVVM 패턴 적용
API를 이용하여 데이터 받아오는 걸 코루틴을 사용하는 방식으로 변경
화면에 새로 불러와져야하는 부분만 작동하도록 상수값을 만들고 조건을 나눔
Home 프래그먼트 MVVM 패턴으로 변경(다이얼로그 프래그먼트 부르는 부분은 아직 못 함)
